src/requirements.txt
django>=4.2
djangorestframework
psycopg2-binary
gunicorn==23.0.0
python-dotenv
django-extensions
Pillow
django-nested-admin==4.1.1
markdown